Bug fixes (#22)
* Fix password reset email.

Places the default sender as the correct email to send the password reset email from. Also places it in the email body.

* Fix wrong password error

Fixes an error (NameError: name 'argon2_exceptions' is not defined) that occurs when entering the correct username but the wrong password.

* Fixed duplicate message

Fixed message "You have no Libraries!
Go to Library Manager and create one." from showing twice when there are no libraries. It would display on the second click of the libraries link.