Fix popups on discover page (#24)
* Added third argument for showDetails() function

Allows for identifying unique DIV ids. Also adjusted the right margin tolerance slightly to avoid some clipping.

* Created unique DIV ids

Added an adendum of unique ids for the latest games and highest rated games rows. This gets passed as a third argument to the showDetails() function and fixes the issue of popping up over a twin cover. The argument is optional and should not break existing function calls.