Check for Future Download File Name
Sometimes the downloads page was being rendered before the zip_game function had time to determine and store the zip_file_location column in the database. This left the file name blank and it did not update with Javascript. Now prior to rendering the template it calculates if the file is going to need to be zipped, and if so determine a path. There may be a more elegant way of doing this, but this was the only consistent way I could find to make sure the file name showed up.