Remove Redundant Code
Since the zip file path is now determined in routes.py before being passed to the zip_game function in utilities.py, it makes sense not to run the same code twice. Now the zip_game function is shortened as it has already been determined if the game needs to be zipped.