Change from Game UUID to Download ID
Update the rows based upon the download ID instead of the game UUID. This allows for multiple upgrade or extra files to be downloaded and updated for a single game. Otherwise it got hung up on being the same game and never updated consequent files. The notify.js file still needs some testing and investigation. It probably should specify which file is available instead of which game.