Add Update Discord Notification
Add function to get the game information using a path (the base path for the update file in this case). Temporarily it is splitting it down to the subdirectory (making the assumption it is in a game's subdirectory). This will change eventually so that main game files are included as well.

Created a function to fire a Discord notification for a new update file.