Remove Duplicate Notifications
Compare the main game update file to the last file being modified. Don't process if they are the same. This keeps multiple notifications from firing when a large file is being overwritten. This only applies to main game files and not updates or extras. It is rest when SharewareZ is rebooted.

Add some debug notifications. Add in all  of the notification options (again?).