Check for Multiple Game Files
Remove debug code from app.py

If this is a main game file update, check to see if there are multiple game files. If there are multiple game files, ignore the update notification. Otherwise if multiple files in a game folder are updated, it will send a notification for each file. Games that are not compressed cannot send update notifications this way.