Add notifications for uncompressed game folders
If the folder contains multiple game files, fire the update notification for only the first file changed. Pass the file name over as the folder name instead of the file the change was detected in. Do not fire any additional update notifications for this game so long as it was the last game updated (resets when a different game is updated and resets when SharewareZ restarts).