Allow Same Game Update
When the game is not compressed, allow that game to have an update notification again even if it was the last game update notification sent. It must however have been at least 60 minutes since the last detection for that game. This places the assumption that an uncompressed game should be fully updated in less than 60 minutes. Covers the rare scenarios where a user is updating just one game over and over again and it is not a compressed game.