File Size Fix and Filtering
Windows has the issue of sometimes firing an OS event for exe file changes when the folder is opened by the user. We now check to see if the exe file is in the main game directory or in the updates or extras directory. If not in the main game directory, we will ignore the modified event as it is not used. If in the main game directory, we now check the size of the game in the DB and compare it to the size of the folder (minutes the updates and extras folders). If the size did not change, ignore the update.