Remove Update and Extras From Main Download
Remove the updates and extras folders from a game when it is zipped. There is some case sensitivity checking here, but if someone names a folder all crazy like "uPdAtEs" and then has the config set as the folder being named "Updates", it is not going to work. For capitalization of first letter it works fine though.