Add Last Updated & More
Remove the update_folder_name from the config file as it is no longer obtained from there.
Create a column in games for last_updated.
Move a whole lot of stuff out of the Discord notifications and into the notification_manager. Files are now processed once and the information can then be sent to various notification services. Also this allows for additional events to occur when games are updated, such as updating the last_updated column in the DB.